innocuous changes suggested from compiler / linter

+++ crates/bevy_gltf/src/loader.rs
@@ -579,8 +579,8 @@ async fn load_texture<'a>(
     let is_srgb = !linear_textures.contains(&gltf_texture.index());
     let mut texture = match gltf_texture.source().source() {
         gltf::image::Source::View { view, mime_type } => {
-            let start = view.offset() as usize;
-            let end = (view.offset() + view.length()) as usize;
+            let start = view.offset();
+            let end = view.offset() + view.length();
             let buffer = &buffer_data[view.buffer().index()][start..end];
             Image::from_buffer(
                 buffer,
